# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRecoveryJobRequest": {
        "additionalProperties": false,
        "description": "Request to recover content for one lecture interval.",
        "properties": {
          "end_ms": {
            "description": "Half-open interval end in ms.",
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "source_event_ids": {
            "description": "Signal events that motivated this request; validated against the session.",
            "items": {
              "type": "string"
            },
            "title": "Source Event Ids",
            "type": "array"
          },
          "start_ms": {
            "description": "Half-open interval start in ms.",
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "start_ms",
          "end_ms"
        ],
        "title": "CreateRecoveryJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobFailure": {
        "description": "A typed, recoverable failure attached to a failed job.",
        "properties": {
          "message": {
            "description": "Human-readable, compassionate failure explanation.",
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/JobFailureReason",
            "description": "Stable typed failure reason."
          }
        },
        "required": [
          "reason",
          "message"
        ],
        "title": "JobFailure",
        "type": "object"
      },
      "JobFailureReason": {
        "description": "Typed, recoverable failure reasons for recovery jobs.",
        "enum": [
          "missing_transcript_context",
          "provider_refused",
          "provider_timeout",
          "provider_malformed_output",
          "validation_failed",
          "internal_error"
        ],
        "title": "JobFailureReason",
        "type": "string"
      },
      "JobStatus": {
        "description": "Lifecycle state of a recovery job.",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "LectureSession": {
        "description": "One running occurrence of a lecture on the shared lecture clock.",
        "properties": {
          "course_id": {
            "description": "Owning course identifier.",
            "title": "Course Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UTC ISO 8601 end time.",
            "title": "Ended At"
          },
          "lecture_id": {
            "description": "Identifier of the underlying lecture record.",
            "title": "Lecture Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/SessionMode",
            "description": "How the session is held."
          },
          "owner_id": {
            "description": "Authenticated user who started the session.",
            "title": "Owner Id",
            "type": "string"
          },
          "session_clock_origin": {
            "description": "UTC ISO 8601 origin of the shared lecture clock.",
            "title": "Session Clock Origin",
            "type": "string"
          },
          "session_id": {
            "description": "Unique identifier for this occurrence.",
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "description": "UTC ISO 8601 start time ending in Z.",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus",
            "description": "Current lifecycle state."
          },
          "title": {
            "description": "Human-readable lecture-session title.",
            "title": "Title",
            "type": "string"
          },
          "zoom_meeting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Zoom meeting identifier when authorized.",
            "title": "Zoom Meeting Id"
          }
        },
        "required": [
          "session_id",
          "lecture_id",
          "owner_id",
          "course_id",
          "title",
          "mode",
          "status",
          "started_at",
          "session_clock_origin"
        ],
        "title": "LectureSession",
        "type": "object"
      },
      "ModelMetadata": {
        "additionalProperties": false,
        "description": "Provider metadata recorded with every generated card.",
        "properties": {
          "data_label": {
            "default": "synthetic",
            "description": "Synthetic for mock usage; measured only for real provider calls.",
            "enum": [
              "synthetic",
              "measured"
            ],
            "title": "Data Label",
            "type": "string"
          },
          "input_character_count": {
            "description": "Bounded input size.",
            "minimum": 0.0,
            "title": "Input Character Count",
            "type": "integer"
          },
          "latency_ms": {
            "description": "Generation latency in milliseconds.",
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "description": "Model identifier used for generation.",
            "title": "Model",
            "type": "string"
          },
          "output_character_count": {
            "description": "Generated output size.",
            "minimum": 0.0,
            "title": "Output Character Count",
            "type": "integer"
          },
          "output_schema_version": {
            "description": "Version of the output schema.",
            "title": "Output Schema Version",
            "type": "string"
          },
          "prompt_version": {
            "description": "Version of the recovery prompt.",
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "description": "Provider name, e.g. ``mock`` or ``openai``.",
            "title": "Provider",
            "type": "string"
          },
          "provider_mode": {
            "description": "Execution mode; mock responses are labeled synthetic.",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Provider Mode",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "provider_mode",
          "prompt_version",
          "output_schema_version",
          "input_character_count",
          "output_character_count",
          "latency_ms"
        ],
        "title": "ModelMetadata",
        "type": "object"
      },
...
      "RecoveryCard": {
        "description": "A grounded, compassionate explanation of missed lecture content.",
        "properties": {
          "card_id": {
            "description": "Unique recovery-card identifier.",
            "title": "Card Id",
            "type": "string"
          },
          "created_at": {
            "description": "UTC ISO 8601 creation time ending in Z.",
            "title": "Created At",
            "type": "string"
          },
          "example_from_lecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Closest lecture excerpt to the missed interval.",
            "title": "Example From Lecture"
          },
          "follow_up_question": {
            "description": "Suggested follow-up for the student.",
            "title": "Follow Up Question",
            "type": "string"
          },
          "key_facts": {
            "description": "Grounded facts from the lecture context.",
            "items": {
              "type": "string"
            },
            "title": "Key Facts",
            "type": "array"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata",
            "description": "Provider and usage metadata."
          },
          "session_id": {
            "description": "Session the card was generated for.",
            "title": "Session Id",
            "type": "string"
          },
          "source_event_ids": {
            "description": "Signal events that motivated this card.",
            "items": {
              "type": "string"
            },
            "title": "Source Event Ids",
            "type": "array"
          },
          "source_timestamps": {
            "description": "Source references and timestamps grounding the card.",
            "items": {
              "$ref": "#/components/schemas/SourceTimestamp"
            },
            "title": "Source Timestamps",
            "type": "array"
          },
          "topic": {
            "description": "Short topic label for what was missed.",
            "title": "Topic",
            "type": "string"
          },
          "what_you_missed": {
            "description": "Compassionate explanation of the missed content.",
            "title": "What You Missed",
            "type": "string"
          }
        },
        "required": [
          "card_id",
          "session_id",
          "source_event_ids",
          "topic",
          "what_you_missed",
          "key_facts",
          "source_timestamps",
          "follow_up_question",
          "model_metadata",
          "created_at"
        ],
        "title": "RecoveryCard",
        "type": "object"
      },
      "RecoveryJob": {
        "description": "Recovery generation job for one requested interval.",
        "properties": {
          "cache_status": {
            "anyOf": [
              {
                "enum": [
                  "miss",
                  "hit"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether generation reused a cached card.",
            "title": "Cache Status"
          },
          "card_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Card reference when completed.",
            "title": "Card Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UTC ISO 8601 completion time ending in Z.",
            "title": "Completed At"
          },
          "created_at": {
            "description": "UTC ISO 8601 creation time ending in Z.",
            "title": "Created At",
            "type": "string"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobFailure"
              },
              {
                "type": "null"
              }
            ],
            "description": "Typed failure when failed."
          },
          "job_id": {
            "description": "Unique recovery-job identifier.",
            "title": "Job Id",
            "type": "string"
          },
          "requested_end_ms": {
            "description": "Requested interval end in ms.",
            "exclusiveMinimum": 0.0,
            "title": "Requested End Ms",
            "type": "integer"
          },
          "requested_start_ms": {
            "description": "Requested interval start in ms.",
            "minimum": 0.0,
            "title": "Requested Start Ms",
            "type": "integer"
          },
          "session_id": {
            "description": "Session the job belongs to.",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current job status."
          }
        },
        "required": [
          "job_id",
          "session_id",
          "status",
          "requested_start_ms",
          "requested_end_ms",
          "created_at"
        ],
        "title": "RecoveryJob",
        "type": "object"
      },
...
      "SessionMode": {
        "description": "How a lecture session is held.",
        "enum": [
          "zoom",
          "in_person"
        ],
        "title": "SessionMode",
        "type": "string"
      },
      "SessionStatus": {
        "description": "Lifecycle state of a lecture session.",
        "enum": [
          "created",
          "active",
          "ending",
          "ended",
          "failed"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
...
      "SourceTimestamp": {
        "additionalProperties": false,
        "description": "A grounded source reference with lecture-relative timestamps.",
        "properties": {
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source transcript chunk.",
            "title": "Chunk Id"
          },
          "end_ms": {
            "description": "Half-open interval end in ms.",
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source system name.",
            "title": "Source"
          },
          "start_ms": {
            "description": "Half-open interval start in ms.",
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          }
        },
        "required": [
          "start_ms",
          "end_ms"
        ],
        "title": "SourceTimestamp",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Typed REST endpoints for lecture sessions, recovery cards, integrations, and anonymous professor summaries.",
    "summary": "Backend API for privacy-first lecture recovery.",
    "title": "Lecture Recovery Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions/{session_id}": {
      "get": {
        "description": "Return the session for an authenticated owner, participant, or professor.",
        "operationId": "read_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve a lecture session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/recovery/cards/{card_id}": {
      "get": {
        "description": "Return one personal recovery card to its owner or the session owner.",
        "operationId": "read_recovery_card_api_v1_sessions__session_id__recovery_cards__card_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve a recovery card",
        "tags": [
          "recovery"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/recovery/jobs": {
      "post": {
        "description": "Recover missed content for ``[start_ms, end_ms)`` in one session.",
        "operationId": "create_recovery_job_api_v1_sessions__session_id__recovery_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecoveryJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request a recovery card for an interval",
        "tags": [
          "recovery"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/recovery/jobs/{job_id}": {
      "get": {
        "description": "Return the job status, card reference, or typed failure.",
        "operationId": "read_recovery_job_api_v1_sessions__session_id__recovery_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve a recovery job",
        "tags": [
          "recovery"
        ]
      }
    },
...
  },
...
}